+++ aikau/src/test/resources/alfresco/upload/UploadTest.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2005-2014 Alfresco Software Limited.
+ * Copyright (C) 2005-2015 Alfresco Software Limited.
  *
  * This file is part of Alfresco
  *
@@ -27,172 +27,152 @@ define(["intern!object",
         "alfresco/TestCommon"], 
         function (registerSuite, expect, assert, require, TestCommon) {
 
-   var testname = "Upload Test";
-         
+   var browser;
    var uploadsSelector = ".alfresco-dialog-AlfDialog .alfresco-upload-AlfUploadDisplay .uploads";
    var successfulUploadsSelector = uploadsSelector + "> .successful table tr";
    var failedUploadsSelector = uploadsSelector + " > .failed table tr";
-
    var aggProgStatusSelector = uploadsSelector + " .aggregate-progress .percentage";
-
    var okButtonSelector = ".alfresco-dialog-AlfDialog .footer > span:first-child > span";
    var cancelButtonSelector = ".alfresco-dialog-AlfDialog .footer > span:nth-child(2) > span";
 
    var dialogDelay = 500;
 
    registerSuite({
-      name: 'Upload Tests',
-      'Upload Failure': function () {
-
-         return TestCommon.loadTestWebScript(this.remote, "/aikau-upload-failure-unit-test", testname)
-
-         // Simulate providing a zero byte file and check the output...
-         .findByCssSelector("#SINGLE_UPLOAD_label")
-            .click()
-            .sleep(dialogDelay)
-         .end()
-
-         .findAllByCssSelector(failedUploadsSelector)
-            .then(function(elements) {
-               TestCommon.log(testname, "Checking that there is a failed upload");
-               assert(elements.length === 1, "Test #1a - Wrong number of failed uploads, expected 1, found: " + elements.length);
-            })
-         .end()
-
-         .findByCssSelector(cancelButtonSelector)
-            .click()
-            .sleep(dialogDelay)
-         .end();
-
+      name: "Upload Failure Tests",
+      setup: function() {
+         browser = this.remote;
+         return TestCommon.loadTestWebScript(this.remote, "/aikau-upload-failure-unit-test", "Upload Failure").end();
       },
-
-      'Bad File Data': function () {
-
-         return TestCommon.loadTestWebScript(this.remote, "/aikau-upload-unit-test", testname)
-
-         // Simulate providing a zero byte file and check the output...
-         .findByCssSelector("#BAD_FILE_DATA_label")
-            .click()
-            .sleep(dialogDelay)
-         .end()
-
-         .findAllByCssSelector(failedUploadsSelector)
-            .then(function(elements) {
-               TestCommon.log(testname, "Checking that there is a failed upload");
-               assert(elements.length === 1, "Test #1a - Wrong number of failed uploads, expected 1, found: " + elements.length);
-            })
-         .end()
-
-         .findByCssSelector(cancelButtonSelector)
-            .click()
-            .sleep(dialogDelay)
-         .end();
-
+      beforeEach: function() {
+         browser.end();
       },
-
-      'Single File Upload': function () {
-
-         var browser = this.remote;
+      teardown: function() {
+         return browser.end().alfPostCoverageResults(browser);
+      },
+      "Upload Failure": function () {
+         // Simulate providing a zero byte file and check the output...
          return browser.findByCssSelector("#SINGLE_UPLOAD_label")
             .click()
             .sleep(dialogDelay)
          .end()
-
          .findAllByCssSelector(failedUploadsSelector)
             .then(function(elements) {
-               TestCommon.log(testname, "Checking that there are no failed uploads");
-               assert(elements.length === 0, "Test #1a - Wrong number of failed uploads, expected 0, found: " + elements.length);
+               assert(elements.length === 1, "Test #1a - Wrong number of failed uploads, expected 1, found: " + elements.length);
             })
          .end()
-
-         .findAllByCssSelector(successfulUploadsSelector)
+         .findByCssSelector(cancelButtonSelector)
+            .click()
+            .sleep(dialogDelay)
+         .end();
+      }
+   });
+   registerSuite({
+      name: "Upload Tests",
+      setup: function() {
+         browser = this.remote;
+         return TestCommon.loadTestWebScript(this.remote, "/aikau-upload-unit-test", "Upload").end();
+      },
+      beforeEach: function() {
+         browser.end();
+      },
+      teardown: function() {
+         return browser.end().alfPostCoverageResults(browser);
+      },
+      "Test bad file data": function () {
+         // Simulate providing a zero byte file and check the output...
+         return browser.findByCssSelector("#BAD_FILE_DATA_label")
+            .click()
+            .sleep(dialogDelay)
+         .end()
+         .findAllByCssSelector(failedUploadsSelector)
             .then(function(elements) {
-               TestCommon.log(testname, "Checking that there is one successful upload");
-               assert(elements.length === 1, "Test #1b - Wrong number of successful uploads, expected 1, found: " + elements.length);
+               assert(elements.length === 1, "Wrong number of failed uploads, expected 1, found: " + elements.length);
             })
          .end()
-
-         .findByCssSelector(aggProgStatusSelector)
+         .findByCssSelector(cancelButtonSelector)
+            .click()
+            .sleep(dialogDelay)
+         .end();
+      },
+      "Test single file upload (no failures)": function () {
+         return browser.findByCssSelector("#SINGLE_UPLOAD_label")
+            .click()
+            .sleep(dialogDelay)
+         .end()
+         .findAllByCssSelector(failedUploadsSelector)
+            .then(function(elements) {
+               assert(elements.length === 0, "Wrong number of failed uploads, expected 0, found: " + elements.length);
+            })
+         .end();
+      },
+      "Test single file upload (one success)": function() {
+         return browser.findAllByCssSelector(successfulUploadsSelector)
+            .then(function(elements) {
+               assert(elements.length === 1, "Wrong number of successful uploads, expected 1, found: " + elements.length);
+            })
+         .end();
+      },
+      "Test single file upload (progress)": function() {
+         return browser.findByCssSelector(aggProgStatusSelector)
             .getVisibleText()
             .then(function(text) {
-               TestCommon.log(testname, "Checking the aggregate progress is 100%");
-               assert(text === "100%", "Test #1b - The aggregate progress was not 100%: " + text);
+               assert(text === "100%", "The aggregate progress was not 100%: " + text);
             })
          .end()
-
          .findByCssSelector(okButtonSelector)
             .click()
             .sleep(dialogDelay)
          .end();
-
       },
-
-      'No Files Upload': function () {
-
-         var browser = this.remote;
+      "Test zero file upload (failed)": function () {
          return browser.findByCssSelector("#NO_FILES_UPLOAD_label")
             .click()
             .sleep(dialogDelay)
          .end()
-
          .findAllByCssSelector(failedUploadsSelector)
             .then(function(elements) {
-               TestCommon.log(testname, "Checking that there are no failed uploads");
-               assert(elements.length === 0, "Test #1a - Wrong number of failed uploads, expected 0, found: " + elements.length);
+               assert(elements.length === 0, "Wrong number of failed uploads, expected 0, found: " + elements.length);
             })
-         .end()
-
-         .findAllByCssSelector(successfulUploadsSelector)
+         .end();
+      },
+      "Test zero file upload (successful)": function() {
+         return browser.findAllByCssSelector(successfulUploadsSelector)
             .then(function(elements) {
-               TestCommon.log(testname, "Checking that there is one successful upload");
-               assert(elements.length === 0, "Test #1b - Wrong number of successful uploads, expected 0, found: " + elements.length);
+               assert(elements.length === 0, "Wrong number of successful uploads, expected 0, found: " + elements.length);
             })
          .end()
-
          .findByCssSelector(okButtonSelector)
             .click()
             .sleep(dialogDelay)
          .end();
-
       },
-
-      'Multi-File Upload': function () {
-
-         var browser = this.remote;
-         
+      "Test Multi-File Upload (failed)": function () {
          return browser.findByCssSelector("#MULTI_UPLOAD_label")
             .click()
             .sleep(dialogDelay)
          .end()
-
          .findAllByCssSelector(failedUploadsSelector)
             .then(function(elements) {
-               TestCommon.log(testname, "Checking that there are no failed uploads");
-               assert(elements.length === 0, "Test #1a - Wrong number of failed uploads, expected 0, found: " + elements.length);
+               assert(elements.length === 0, "Wrong number of failed uploads, expected 0, found: " + elements.length);
             })
-         .end()
-
-         .findAllByCssSelector(successfulUploadsSelector)
+         .end();
+      },
+      "Test Multi-File Upload (successful)": function () {
+         return browser.findAllByCssSelector(successfulUploadsSelector)
             .then(function(elements) {
-               TestCommon.log(testname, "Checking that there is one successful upload");
-               assert(elements.length === 4, "Test #1b - Wrong number of successful uploads, expected 4, found: " + elements.length);
+               assert(elements.length === 4, "Wrong number of successful uploads, expected 4, found: " + elements.length);
             })
          .end()
-
          .findByCssSelector(aggProgStatusSelector)
             .getVisibleText()
             .then(function(text) {
-               TestCommon.log(testname, "Checking the aggregate progress is 100%");
-               assert(text === "100%", "Test #1b - The aggregate progress was not 100%: " + text);
+               assert(text === "100%", "The aggregate progress was not 100%: " + text);
             })
          .end()
-
          .findByCssSelector(okButtonSelector)
             .click()
             .sleep(dialogDelay)
-         .end()
-
-         .alfPostCoverageResults(browser);
+         .end();
       }
    });
 });
